+++ odt2pdf/0.5.0.3/HISTORY.md
@@ -27,3 +27,5 @@
 - GrahaOdt2HtmlConverter 에서 ```div#GrahaOdt2HtmlConverterWrapper``` 를 document.body 에 먼저 추가하는 것으로 변경
 	- ```div#GrahaOdt2HtmlConverterWrapper``` 가 렌더링 되지 않았다면, data-anchor-type 이 page 인 것들이 자리를 제대로 잡지 못하는 버그가 있다,
 - 기능상의 변화나 버그의 수정없이 일부 함수의 구현을 GrahaOdtPageSplitterUtility.js 로 이동함.
+- heightForParent 함수가 td 대신 tr 의 height 를 리턴하는 경우(IE 11 에서만 발생하는 것으로 추정) td 의 padding-top 과 padding-bottom 을 공제하도록 변경(버그 fix 이지만, IE 11 관련이므로, 버전을 변경하지 않음)
+- window.devicePixelRatio 를 1.25 로 고정함(잠재적인 것이므로, 버전을 변경하지 않음)